@@ -6,7 +6,7 @@ export interface GitHubRelease {
 }
 
 export const getRelease = async (version: string): Promise<GitHubRelease | null> => {
-  const url: string = `https://github.com/containerd/containerd/releases/${version}`;
+  const url = `https://github.com/containerd/containerd/releases/${version}`;
   const http: httpm.HttpClient = new httpm.HttpClient('ghaction-setup-containerd');
   return (await http.getJson<GitHubRelease>(url)).result;
 };
